third_party/blink/web_tests/fast/performance/performance-observer-expected.txt

This tests that PerformanceObserver gets called the right number of times.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS entries[0].entryType is expectedEntries[0].entryType
PASS entries[0].name is expectedEntries[0].name
PASS entries[1].entryType is expectedEntries[1].entryType
PASS entries[1].name is expectedEntries[1].name
PASS entries[2].entryType is expectedEntries[2].entryType
PASS entries[2].name is expectedEntries[2].name
PASS successfullyParsed is true

TEST COMPLETE
